+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Changed
 - IndexEntryException has been replaced by IndexElementException
+- Guzzle 6 is now used to get page content of elements instead of the Twig template renderer
 
 ## 1.0.0 - 2018-12-12
 ### Added